PMM-4192 Removed redundant check
Removed MySQL flavor comparisson for MySQL 8 because there is only one
vendor for v8.0
Also updated the way we get the vendor name from mysql version_comment
since not all vendors and versions have the same capitalizacion in the
name. Now the name is being converted to lowercase before comparing.